modified: split init
