+++ philo/src/init/init_data.c
@@ -12,7 +12,7 @@
 
 #include "../../include/philo.h"
 
-t_bool	is_invalid_input(char *nbrp)
+static t_bool	is_invalid_input(char *nbrp)
 {
 	long long	lln;
 	size_t		i;
@@ -36,7 +36,7 @@ t_bool	is_invalid_input(char *nbrp)
 	return (FALSE);
 }
 
-t_data	*init_num_datas(t_data *data, char **argv)
+static t_data	*init_num_datas(t_data *data, char **argv)
 {
 	if (is_invalid_input(argv[1]))
 		return (init_error(data));
@@ -54,91 +54,7 @@ t_data	*init_num_datas(t_data *data, char **argv)
 	return (data);
 }
 
-t_philo	*init_philo(t_data *data, size_t i)
-{
-	t_philo	*philo;
-
-	philo = (t_philo *)malloc(sizeof(t_philo));
-	if (!philo)
-		return (NULL);
-	memset(philo, 0, sizeof(t_philo));
-	philo->id = i;
-	philo->right = data->forks[i];
-	if (i == 0)
-		philo->left = data->forks[data->num_of_philo - 1];
-	else
-		philo->left = data->forks[i - 1];
-	philo->eat_count = 0;
-	philo->data = data;
-	philo->count_mtx = &(data->count_mtxs[i]);
-	philo->msg_mtx = &(data->msg_mtx);
-	philo->time_to_die = data->time_to_die;
-	philo->time_to_eat = data->time_to_eat;
-	philo->time_to_sleep = data->time_to_sleep;
-	philo->must_eat = data->must_eat;
-	return (philo);
-}
-
-t_data *init_philos(t_data *data)
-{
-	size_t	i;
-
-	data->philos = (t_philo **)malloc(sizeof(t_philo *) * data->num_of_philo);
-	if (!(data->philos))
-		return (init_error(data));
-	memset(data->philos, 0, sizeof(t_philo *) * data->num_of_philo);
-	i = 0;
-	while (i < (size_t)data->num_of_philo)
-	{
-		data->philos[i] = init_philo(data, i);
-		if (!(data->philos[i]))
-			return (init_error(data));
-		i++;
-	}
-	return (data);
-}
-
-t_fork	*init_fork(size_t i)
-{
-	t_fork	*fork;
-
-	fork = (t_fork *)malloc(sizeof(t_fork));
-	if (!fork)
-		return (NULL);
-	if (pthread_mutex_init(&(fork->mtx), NULL))
-		return (NULL);
-	fork->id = i;
-	return (fork);
-}
-
-t_data	*init_forks(t_data *data)
-{
-	size_t	i;
-
-	data->forks = (t_fork **)malloc(sizeof(t_fork *) * data->num_of_philo);
-	if (!(data->forks))
-		return (init_error(data));
-	memset(data->forks, 0, sizeof(t_fork *) * data->num_of_philo);
-	i = 0;
-	while (i < (size_t)data->num_of_philo)
-	{
-		data->forks[i] = init_fork(i);
-		if (!(data->forks[i]))
-			return (init_error(data));
-		i++;
-	}
-	return (data);
-}
-
-t_data	*init_threads(t_data *data)
-{
-	data->threads = (pthread_t *)malloc(sizeof(pthread_t) * data->num_of_philo);
-	if (!(data->threads))
-		return (init_error(data));
-	return (data);
-}
-
-t_data	*init_must_eat(t_data *data, int argc, char *argv[])
+static t_data	*init_must_eat(t_data *data, int argc, char *argv[])
 {
 	if (argc != 6)
 		return (data);
@@ -148,33 +64,10 @@ t_data	*init_must_eat(t_data *data, int argc, char *argv[])
 	return (data);
 }
 
-t_data	*init_count_mtxs(t_data *data)
+static t_data	*init_threads(t_data *data)
 {
-	pthread_mutex_t	*count_mtxs;
-	size_t			i;
-
-	count_mtxs = (pthread_mutex_t *)malloc(sizeof(pthread_mutex_t) * data->num_of_philo);
-	if (!count_mtxs)
-		return (init_error(data));
-	i = 0;
-	while (i < data->num_of_philo)
-	{
-		if (pthread_mutex_init(&(data->count_mtxs[i]), NULL))
-			return (init_error(data));
-		i++;
-	}
-	return (data);
-}
-
-t_data	*init_mutexs(t_data *data)
-{
-	data = init_count_mtxs(data);
-	if (!data)
-		return (init_error(data));
-	data = init_forks(data);
-	if (!data)
-		return (init_error(data));
-	if (pthread_mutex_init(&(data->msg_mtx), NULL))
+	data->threads = (pthread_t *)malloc(sizeof(pthread_t) * data->num_of_philo);
+	if (!(data->threads))
 		return (init_error(data));
 	return (data);
 }
